Move transport/connection.js to upper level Since it is used across several modules. Hence WebSocketConnection should be moved into a separate file.
